+++ docs/source/changes.rst
@@ -21,6 +21,8 @@ Improvements
 ------------
 - :func:`hydra_zen.launch` now supports dictionary overrides. See :pull:`313`.
 - :class:`hydra_zen.ZenStore` now provides specialized support for storing instances/subclasses of `HydraConf`. See :issue:`395`.
+- Adds auto-config support for jax 0.4.0. See :pull:`414`.
+- Improved the type annotations of :class:`~hydra_zen.wrappers.ZenStore`. See :pull:`409`.
 
 
 --------------------------